catching/handling psql errors correctly

+++ backend/middleware/songrequest.go
@@ -4,15 +4,15 @@ import (
 	"backend/twitch-bot/api"
 	"backend/twitch-bot/database"
 	"backend/twitch-bot/models"
-	"runtime"
+	"log"
 	"strconv"
-	"strings"
 
 	"github.com/gofiber/fiber/v2"
 )
 
+// Middleware SongRequest
 func SongRequest(c *fiber.Ctx) error {
-	runtime.GOMAXPROCS(4)
+	// Creates a Query Struct for the query parameters the GET request will take in
 	type Query struct {
 		Channel string `query:"channel"`
 		User    string `query:"user"`
@@ -25,6 +25,7 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": err,
 		})
 	}
+	// Checks if query is empty if it is then return back to the request that the query is missing
 	if query.Q == "" {
 		clientData := models.ClientData{
 			Status:  "fail",
@@ -35,7 +36,7 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-
+	// Checks if user is empty if it is then return back to the request that the user is missing
 	if query.User == "" {
 		clientData := models.ClientData{
 			Status:  "fail",
@@ -46,7 +47,7 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-
+	// Checks if channel is empty if it is then return back to the request that the channel is missing
 	if query.Channel == "" {
 		clientData := models.ClientData{
 			Status:  "fail",
@@ -57,6 +58,7 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
+	// Gets the songData from the youtube api using the query q(uery) from the request
 	songData := api.GetSongFromSearch(query.Q)
 	if songData.PageInfo.TotalResults == 0 {
 		clientData := models.ClientData{
@@ -68,8 +70,7 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-	songData.Items[0].Snippet.Title = strings.ReplaceAll(songData.Items[0].Snippet.Title, "&amp;", "&")
-	//songData.Items[0].Snippet.Title = strings.ReplaceAll(songData.Items[0].Snippet.Title, "&#39;", "'")
+	// Gets the duration of the video using the videoID, we have to make separate API calls here because the search api doesn't return the video duration
 	songDuration := api.GetVideoDuration(songData.Items[0].ID.VideoID)
 	if songDuration.IsLiveStream {
 		clientData := models.ClientData{
@@ -91,9 +92,20 @@ func SongRequest(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-	db := database.InitializeConnection()
-	database.CreateTable(query.Channel, db)
-	latestSongPos := database.GetLatestSongPosition(db, query.Channel)
+	// Makes the initial DB connection and attempts to create the table
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(map[string]interface{}{
+			"error": dbConnErr.Error(),
+		})
+	}
+	err := database.CreateTable(query.Channel, db)
+	if err != nil {
+		log.Fatalln(err)
+	}
+	// Attempts to get the latestSongPosition
+	latestSongPos, err := database.GetLatestSongPosition(db, query.Channel)
+
 	if latestSongPos >= 20 {
 		clientData := models.ClientData{
 			Status:  "fail",
@@ -114,11 +126,20 @@ func SongRequest(c *fiber.Ctx) error {
 		Position: latestSongPos + 1,
 	}
 
+	// if the table is being created for the first time, the GetLatestSongPosition function can't query through because it thinks that the table was never created so it throws a pq error of undefined_table
+	// so we catch this error and if we do get the "undefined_table" error then create the table "again"(even though it was never created) then insert it
+	if err != nil {
+		err := database.CreateTable(query.Channel, db)
+		if err != nil {
+			c.Next()
+		}
+	}
+
 	dataError := database.InsertSong(db, song, query.Channel)
-	if dataError != "" {
+	if dataError != nil {
 		clientData := models.ClientData{
 			Status:  "fail",
-			Message: dataError,
+			Message: dataError.Error(),
 			Data:    nil,
 		}
 		return c.JSON(map[string]interface{}{
@@ -137,16 +158,17 @@ func SongRequest(c *fiber.Ctx) error {
 	return c.JSON(clientData)
 }
 
+
+
+
+
+
+// Middleware function that returns all the songs in that current table.
 func FetchAllSongs(c *fiber.Ctx) error {
 	type Query struct {
 		Channel string `query:"channel"`
 	}
 	query := new(Query)
-	if err := c.QueryParser(query); err != nil {
-		return c.JSON(&fiber.Map{
-			"error": err,
-		})
-	}
 	if query.Channel == "" {
 		clientData := models.ClientData{
 			Status:  "fail",
@@ -157,7 +179,12 @@ func FetchAllSongs(c *fiber.Ctx) error {
 			"error": clientData.Message,
 		})
 	}
-	db := database.InitializeConnection()
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(map[string]interface{}{
+			"error": dbConnErr.Error(),
+		})
+	}
 	songs, err := database.GetAllSongRequests(query.Channel, db)
 	if err != nil {
 		return c.JSON(fiber.Map{
@@ -169,6 +196,7 @@ func FetchAllSongs(c *fiber.Ctx) error {
 	})
 }
 
+// Middleware function that deletes the song based of the id passed in the request query.
 func DeleteSong(c *fiber.Ctx) error {
 	type Query struct {
 		Channel string `query:"channel"`
@@ -190,7 +218,12 @@ func DeleteSong(c *fiber.Ctx) error {
 			"error": "missing channel to delete the song from",
 		})
 	}
-	db := database.InitializeConnection()
+	db, dbConnErr := database.InitializeConnection()
+	if dbConnErr != nil {
+		return c.JSON(&fiber.Map{
+			"error": dbConnErr.Error(),
+		})
+	}
 	err := database.DeleteSong(q.Channel, q.Id, db)
 	if err != nil {
 		return c.JSON(&fiber.Map{
